week2-docker/flask-postgres-redis-app/requirements.txt
flask==2.3.0
redis==4.5.0
psycopg2-binary==2.9.6
requests==2.31.0
